move unused and unneeded isRTL function to dev site package
