--- packages/fast-web-utilities/src/localization.spec.ts
@@ -1,11 +1,0 @@
-import { isRTL } from "./localization";
-
-describe("Check to see if a locale is RTL", (): void => {
-    test("should return false if locale is not valid", (): void => {
-        expect(isRTL("")).toBe(false);
-    });
-
-    test("en", (): void => {
-        expect(isRTL("en")).toBe(false);
-    });
-});
